requirements.txt
streamlit
matplotlib==3.5.1
lxml
numpy==1.21.5
pandas==1.4.2
scikit-image 
pillow
imgaug==0.4.0
scikit-learn==1.0.2
keras
tensorflow
cv2
opencv-python-headless
requests
beautifulsoup4